+++ Asteroids++/FileMenager.cpp
@@ -52,10 +52,10 @@ void FileMenager::setDataFromFile() const {
 
 	file.close();
 
-	FileMenager::screenData.size_height = dataMap["screen_size_height"];
-	FileMenager::screenData.size_width = dataMap["screen_size_width"];
+	FileMenager::screenData.size_height = static_cast<unsigned int>(dataMap["screen_size_height"]);
+	FileMenager::screenData.size_width = static_cast<unsigned int>(dataMap["screen_size_width"]);
 
-	FileMenager::playerData.size = dataMap["player_size"];
+	FileMenager::playerData.size = static_cast<int>(dataMap["player_size"]);
 	FileMenager::playerData.start_position_x = dataMap["player_start_position_x"];
 	FileMenager::playerData.start_position_y = dataMap["player_start_position_y"];
 	FileMenager::playerData.start_position_angle = dataMap["player_start_position_angle"];
@@ -64,7 +64,7 @@ void FileMenager::setDataFromFile() const {
 	FileMenager::playerData.bullet_speed = dataMap["player_bullet_speed"];
 	FileMenager::playerData.bullet_shoot_delay = dataMap["player_bullet_shoot_delay"];
 	FileMenager::playerData.bullet_lifetime = dataMap["player_bullet_lifetime"];
-	FileMenager::playerData.bullet_size = dataMap["player_bullet_size"];
+	FileMenager::playerData.bullet_size = static_cast<int>(dataMap["player_bullet_size"]);
 	FileMenager::playerData.sprite_cycle_time = dataMap["player_sprite_cycle_time"];
 	FileMenager::playerData.dash_duration = dataMap["player_dash_duration"];
 	FileMenager::playerData.dash_length = static_cast<int>(dataMap["player_dash_length"]);
